db module cleanup, u32 -> usize to avoid potential panic, fix exit 1 condition in keycloak container health check

+++ backend/src/db/table.rs
@@ -10,9 +10,13 @@ pub(crate) trait TableRow<PrimaryKey> {
 
 // Table can be reimplemented with associated types for the same reason.
 pub(crate) trait Table<PrimaryKey, Row> where Row: TableRow<PrimaryKey> {
+
+    /// Insert one or more rows into the table.
     fn insert(&mut self, row: Vec<Row>) -> Result<Vec<PrimaryKey>, String>;
 
+    /// Get a row from the table by its primary key.
     fn get(&self, key: &PrimaryKey) -> Result<Row, String>;
 
-    fn list(&self, limit: u32) -> Result<Vec<Row>, String>;
+    /// List all rows from the table, up to some limit.
+    fn list(&self, limit: usize) -> Result<Vec<Row>, String>;
 }
