+++ backend/src/db/tables/posts_by_id.rs
@@ -35,5 +35,5 @@ impl From<Post> for PostsByIdTableRow {
 pub(crate) trait PostsByIdTableLike: Sync + Send {
     fn insert(&mut self, row: Vec<PostsByIdTableRow>) -> Result<Vec<Uuid>, String>;
     fn get(&self, key: &Uuid) -> Result<PostsByIdTableRow, String>;
-    fn list(&self, limit: u32) -> Result<Vec<PostsByIdTableRow>, String>;
+    fn list(&self, limit: usize) -> Result<Vec<PostsByIdTableRow>, String>;
 }
